@@ -1,16 +1,14 @@
 ﻿using Android.App;
 using Android.Runtime;
 
-namespace FinanMan.Maui
+namespace FinanMan.Maui;
+[Application]
+public class MainApplication : MauiApplication
 {
-    [Application]
-    public class MainApplication : MauiApplication
+    public MainApplication(IntPtr handle, JniHandleOwnership ownership)
+        : base(handle, ownership)
     {
-        public MainApplication(IntPtr handle, JniHandleOwnership ownership)
-            : base(handle, ownership)
-        {
-        }
-
-        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
     }
+
+    protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
 }
