@@ -1,16 +1,14 @@
 ﻿using ObjCRuntime;
 using UIKit;
 
-namespace FinanMan.Maui
+namespace FinanMan.Maui;
+public class Program
 {
-    public class Program
+    // This is the main entry point of the application.
+    static void Main(string[] args)
     {
-        // This is the main entry point of the application.
-        static void Main(string[] args)
-        {
-            // if you want to use a different Application Delegate class from "AppDelegate"
-            // you can specify it here.
-            UIApplication.Main(args, null, typeof(AppDelegate));
-        }
+        // if you want to use a different Application Delegate class from "AppDelegate"
+        // you can specify it here.
+        UIApplication.Main(args, null, typeof(AppDelegate));
     }
 }
