@@ -2,7 +2,7 @@
 @using FinanMan.Shared.DataEntryModels
 
 <section class="transaction-history">
-    <table>
+    <table id="tblTransactionHistory-@_id" class="data-grid">
         <colgroup id="pendingCol" />
         <colgroup id="transDateCol" />
         <colgroup id="accountCol" />
@@ -46,7 +46,6 @@
                                     <li class="divider"><hr /></li>
                                     <li class="menu-item">Filter</li>
                                     <li class="menu-item">Clear Filter</li>
-
                                 </ul>
                             </div>
                         </div>
@@ -60,13 +59,13 @@
             @if (SortedTransactions is null)
             {
                 <tr>
-                    <td colspan="6">@Localizer["Loading"]</td>
+                    <td colspan="6">@SharedLocalizer["Loading"]</td>
                 </tr>
             }
             else if (!SortedTransactions.Any())
             {
                 <tr>
-                    <td colspan="6">@Localizer["NoTransactionsMessage"]</td>
+                    <td colspan="6">@SharedLocalizer["NoTransactionsMessage"]</td>
                 </tr>
             }
             else
