@@ -1,6 +1,7 @@
 ﻿@page "/data-entry"
 @page "/data-entry/{EntType}"
 
+@using FinanMan.BlazorUi.Components.AccountSummaryComponents
 @using FinanMan.BlazorUi.Components.DataEntryComponents
 @using FinanMan.Database.Models.Tables;
 @using FinanMan.Shared.General;
@@ -10,9 +11,9 @@
 <div id="dataEntryPage" class="page-content">
     <header>
         <ul class="data-entry-tabs">
-            <li><button type="button" class="tab-button invert@(_entryType == EntryType.Payments ? " active" : string.Empty)" @onclick="_ => HandleEntryTabClicked(EntryType.Payments)">Payments</button></li>
-            <li><button type="button" class="tab-button invert@(_entryType == EntryType.Deposits ? " active" : string.Empty)" @onclick="_ => HandleEntryTabClicked(EntryType.Deposits)">Deposits</button></li>
-            <li><button type="button" class="tab-button invert@(_entryType == EntryType.Transfers ? " active" : string.Empty)" @onclick="_ => HandleEntryTabClicked(EntryType.Transfers)">Transfers</button></li>
+            <li><button type="button" class="tab-button invert@(_entryType == EntryType.Payments ? " active" : string.Empty)" @onclick="_ => HandleEntryTabClicked(EntryType.Payments)">@SharedLocalizer["Payments"]</button></li>
+            <li><button type="button" class="tab-button invert@(_entryType == EntryType.Deposits ? " active" : string.Empty)" @onclick="_ => HandleEntryTabClicked(EntryType.Deposits)">@SharedLocalizer["Deposits"]</button></li>
+            <li><button type="button" class="tab-button invert@(_entryType == EntryType.Transfers ? " active" : string.Empty)" @onclick="_ => HandleEntryTabClicked(EntryType.Transfers)">@SharedLocalizer["Transfers"]</button></li>
         </ul>
     </header>
     <section class="data-entry-section">
@@ -31,6 +32,7 @@
                 break;
         }
     </section>
+    <AccountSummary />
     <TransactionHistory />
     @*<button type="button" @onclick="HandleCreateLookupClicked">Hey look, a button!</button>*@
     @if(_newLookupItemResp?.WasError ?? false)
