@@ -1,5 +1,7 @@
 ﻿@page "/transactions"
 
-<h3>Transaction History</h3>
+<section class="transaction-history-wrapper">
+    <h3>@SharedLocalizer["TransactionHistory"]</h3>
 
-<TransactionHistoryGrid />
+    <TransactionHistoryGrid />
+</section>
