@@ -64,8 +64,8 @@ namespace FinanMan.Database.Migrations.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<double>("Amount")
-                        .HasColumnType("float");
+                    b.Property<decimal>("Amount")
+                        .HasColumnType("decimal(18,2)");
 
                     b.Property<int>("DepositReasonId")
                         .HasColumnType("int");
@@ -94,6 +94,9 @@ namespace FinanMan.Database.Migrations.Migrations
                     b.Property<bool>("Deleted")
                         .HasColumnType("bit");
 
+                    b.Property<bool>("IncreaseOnPayment")
+                        .HasColumnType("bit");
+
                     b.Property<DateTime>("LastUpdated")
                         .HasColumnType("datetime2");
 
@@ -116,6 +119,7 @@ namespace FinanMan.Database.Migrations.Migrations
                         {
                             Id = 1,
                             Deleted = false,
+                            IncreaseOnPayment = false,
                             LastUpdated = new DateTime(2022, 7, 25, 17, 41, 0, 0, DateTimeKind.Unspecified),
                             Name = "Checking",
                             SortOrder = 1
@@ -124,6 +128,7 @@ namespace FinanMan.Database.Migrations.Migrations
                         {
                             Id = 2,
                             Deleted = false,
+                            IncreaseOnPayment = false,
                             LastUpdated = new DateTime(2022, 7, 25, 17, 41, 0, 0, DateTimeKind.Unspecified),
                             Name = "Savings",
                             SortOrder = 2
@@ -132,6 +137,7 @@ namespace FinanMan.Database.Migrations.Migrations
                         {
                             Id = 3,
                             Deleted = false,
+                            IncreaseOnPayment = false,
                             LastUpdated = new DateTime(2022, 7, 25, 17, 41, 0, 0, DateTimeKind.Unspecified),
                             Name = "Credit Card",
                             SortOrder = 3
@@ -140,6 +146,7 @@ namespace FinanMan.Database.Migrations.Migrations
                         {
                             Id = 4,
                             Deleted = false,
+                            IncreaseOnPayment = false,
                             LastUpdated = new DateTime(2022, 7, 25, 17, 41, 0, 0, DateTimeKind.Unspecified),
                             Name = "Cash",
                             SortOrder = 4
@@ -541,8 +548,8 @@ namespace FinanMan.Database.Migrations.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<double>("Amount")
-                        .HasColumnType("float");
+                    b.Property<decimal>("Amount")
+                        .HasColumnType("decimal(18,2)");
 
                     b.Property<string>("Detail")
                         .HasMaxLength(255)
@@ -640,8 +647,8 @@ namespace FinanMan.Database.Migrations.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<double>("Amount")
-                        .HasColumnType("float");
+                    b.Property<decimal>("Amount")
+                        .HasColumnType("decimal(18,2)");
 
                     b.Property<int>("TargetAccountId")
                         .HasColumnType("int");
