@@ -6,7 +6,7 @@ public partial class PaymentDetail
     public int PaymentId { get; set; }
     public int LineItemTypeId { get; set; }
     public string? Detail { get; set; }
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
 
     public virtual Payment Payment { get; set; } = default!;
     public virtual LuLineItemType LineItemType { get; set; } = default!;
