@@ -5,7 +5,7 @@ public partial class Transfer
     public int Id { get; set; }
     public int TransactionId { get; set; }
     public int TargetAccountId { get; set; }
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
 
     public virtual Transaction Transaction { get; set; } = default!;
     public virtual Account TargetAccount { get; set; } = default!;
