@@ -1,17 +1,14 @@
 ﻿using FinanMan.BlazorUi.Extensions;
 using FinanMan.Shared.Extensions;
 using FinanMan.SharedClient.Extensions;
+using Microsoft.Extensions.Logging;
 
 namespace FinanMan.Maui;
-
 public static class MauiProgram
 {
     public static MauiApp CreateMauiApp()
     {
         var builder = MauiApp.CreateBuilder();
-
-        builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri("https://localhost:7176/") });
-
         builder
             .UseMauiApp<App>()
             .ConfigureFonts(fonts =>
@@ -20,13 +17,25 @@ public static class MauiProgram
             });
 
         builder.Services.AddMauiBlazorWebView();
+
 #if DEBUG
         builder.Services.AddBlazorWebViewDeveloperTools();
+        builder.Logging.AddDebug();
 #endif
+
+        var baseAddress = "https://localhost:7176";
+
+#if !DEBUG 
+        // TODO: Overwrite the baseAddress with the address of the server
+#endif
+
+        builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(baseAddress) });
+
 
         builder.Services.AddFinanManLocalization();
         builder.Services.AddStateManagement();
         builder.Services.AddClientServices();
+
 
         return builder.Build();
     }
