@@ -24,12 +24,17 @@ async function onInstall(event) {
 
 async function onActivate(event) {
     console.info('Service worker: Activate');
-
+    
     // Delete unused caches
     const cacheKeys = await caches.keys();
     await Promise.all(cacheKeys
         .filter(key => key.startsWith(cacheNamePrefix) && key !== cacheName)
         .map(key => caches.delete(key)));
+}
+
+function onInstall(event) {
+    self.skipWaiting();
+    console.info("installed without waiting");
 }
 
 async function onFetch(event) {
